@Auth ,@AuthUser  권한

+++ src/main/java/com/bit2016/mysite/controller/BoardController.java
@@ -2,8 +2,6 @@ package com.bit2016.mysite.controller;
 
 import java.util.HashMap;
 
-import javax.servlet.http.HttpSession;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -15,6 +13,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.bit2016.mysite.service.BoardService;
 import com.bit2016.mysite.vo.Board;
 import com.bit2016.mysite.vo.Users;
+import com.bit2016.security.Auth;
+import com.bit2016.security.AuthUser;
 
 @Controller
 @RequestMapping("/board")
@@ -40,15 +40,17 @@ public class BoardController {
 		return "board/view";
 	}
 
+	@Auth
 	@RequestMapping(value = "/write", method = RequestMethod.POST)
 	public String write(@ModelAttribute Board vo, @RequestParam(value = "p", required = true, defaultValue = "1") int p,
-			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, HttpSession session,
+			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, @AuthUser Users authUser,
 			Model model) {
-		boardService.write(vo, session);
+		boardService.write(vo, authUser);
 		return "redirect:/board?p=" + p + "&kwd=" + kwd;
 
 	}
 
+	@Auth
 	@RequestMapping(value = "/write", method = RequestMethod.GET)
 	public String write(@RequestParam(value = "p", required = true, defaultValue = "1") int p,
 			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, Model model) {
@@ -57,57 +59,74 @@ public class BoardController {
 		return "board/write";
 	}
 
+	@Auth
 	@RequestMapping("/delete")
 	public String delete(@RequestParam(value = "no", required = true, defaultValue = "1") Long no,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
-			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, HttpSession session) {
-		Users authUser = (Users) session.getAttribute("authUser");
+			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, @AuthUser Users authUser) {
 		Board board = boardService.view(no);
-
-		if (authUser != null) {
+		if (authUser != null && board != null) {
 			if (board.getUserNo() == authUser.getNo()) {
 				boardService.delete(no);
 			} else {
-				return "redirect:/user/loginform";
+				return "redirect:/board?p=" + p;
 			}
+		} else if (authUser != null) {
+			return "redirect:/board?p=" + p;
 		} else {
-			return "redirect:/user/loginform";
+			return "redirect:/board?p=" + p;
 		}
-
-		return "redirect:/board?p=" + p + "&kwd=" + 1;
+		return "redirect:/board?p=" + p + "&kwd=" + "";
 	}
 
+	@Auth
 	@RequestMapping(value = "/modify", method = RequestMethod.GET)
 	public String modify(@RequestParam(value = "no", required = true, defaultValue = "1") Long no,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
-			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, Model model) {
-		HashMap<String, Object> hm = boardService.view(no, p);
-		model.addAttribute("hm", hm);
-		model.addAttribute("kwd", kwd);
-		return "board/modify";
+			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, @AuthUser Users authUser,
+			Model model) {
+		Board board = boardService.view(no);
+		if (authUser != null && board != null) {
+			if (board.getUserNo() == authUser.getNo()) {
+				HashMap<String, Object> hm = boardService.view(no, p);
+				model.addAttribute("hm", hm);
+				model.addAttribute("kwd", kwd);
+				return "board/modify";
+			} else {
+				return "redirect:/board/view?no=" + no + "&p=" + p + "&kwd=" + kwd;
+			}
+		} else if (authUser != null) {
+			return "redirect:/board/view?no=" + no + "&p=" + p + "&kwd=" + kwd;
+		} else {
+			return "redirect:/loginform";
+		}
+
 	}
 
+	@Auth
 	@RequestMapping(value = "/modify", method = RequestMethod.POST)
 	public String modify(@ModelAttribute Board vo,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
 			@RequestParam(value = "no", required = true, defaultValue = "1") Long no,
-			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, HttpSession session) {
-		Users authUser = (Users) session.getAttribute("authUser");
+			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, @AuthUser Users authUser) {
 		Board board = boardService.view(no);
-		if (authUser != null) {
+		if (authUser != null && board != null) {
 			if (board.getUserNo() == authUser.getNo()) {
 				boardService.update(vo);
 			} else {
-				return "redirect:/user/loginform";
+				return "redirect:/board/view?no=" + no + "&p=" + p + "&kwd=" + kwd;
 			}
+		} else if (authUser != null) {
+			return "redirect:/board/view?no=" + no + "&p=" + p + "&kwd=" + kwd;
 		} else {
-			return "redirect:/user/loginform";
+			return "redirect:/loginform";
 		}
 
 		return "redirect:/board/view?no=" + no + "&p=" + p + "&kwd=" + kwd;
 
 	}
 
+	@Auth
 	@RequestMapping(value = "/reply", method = RequestMethod.GET)
 	public String reply(@RequestParam(value = "rno", required = true, defaultValue = "0") Long rno,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
@@ -118,12 +137,13 @@ public class BoardController {
 		return "board/reply";
 	}
 
+	@Auth
 	@RequestMapping(value = "/reply", method = RequestMethod.POST)
 	public String reply(@RequestParam(value = "rno", required = true, defaultValue = "0") Long rno,
 			@RequestParam(value = "p", required = true, defaultValue = "1") int p,
 			@RequestParam(value = "kwd", required = true, defaultValue = "") String kwd, @ModelAttribute Board vo,
-			HttpSession session, Model model) {
-		boardService.reply(vo, session, rno);
+			@AuthUser Users authUser, Model model) {
+		boardService.reply(vo, authUser, rno);
 		return "redirect:/board?p=" + p + "&kwd=" + kwd;
 	}
 }
